fix(scripts): Fix PopulateDropdown script
Because assets not in the scene are not shipped with the game, I tried
moving assets I wanted to load dynamically into a folder called
Resources, and using Resource.Load(). However, that doesn't seem to
work. I've fallen back to hardcoding the furniture categories until
a better solution can be found for both Editor and Android simulation.